Working to Redesign the application with States
Going attempt changing from my original idea of a login screen, then access to the web application to the user having full access to the web application first! And once they have everything the way they want it to be. The user can then enter the data for the email and that will be sent to the specified receipt. Additionally if the user wants to see previous postcards, if they are still logged in then they can simply select the history option. However, if it's from a previous session then they'll have to log in and the server will verify by sending a notification email to the user. Which'll then give access to the user to look at their history.